refactor: remanes index services
