+++ src/services/app/unit-measure/index.ts
@@ -1,7 +1,1 @@
-import { endpoints } from '@src/services/config/endpoints';
-import { UnitMeasure } from '@src/store/slices/unit-measure/types';
-import { httpGET } from '../../http-client';
-
-export function getAllUnitMeasures(): Promise<UnitMeasure[]> {
-  return httpGET<UnitMeasure[]>(endpoints.unitMeasure.GET_ALL);
-}
+export { getAllUnitMeasures } from './unit-measure';
