+++ src/services/auth/index.ts
@@ -1,2 +1,11 @@
-export * from './sign-up';
-export * from './sign-in';
+export type { RegisterProducerDTO } from './sign-up';
+export type { Credentials } from './sign-in';
+
+export { signIn } from './sign-in';
+export {
+  registerConsumer,
+  registerProducer,
+  reserveDocument,
+  reserveEmail,
+  reservePhone,
+} from './sign-up';
